+++ src/extension.ts
@@ -1,5 +1,6 @@
 import * as vscode from "vscode";
 import { MezDefinitionProvider } from "./definitionProvider";
+import { MezReferenceProvider } from "./referenceProvider";
 import { MezSymbolIndex } from "./symbolIndex";
 
 let index: MezSymbolIndex | undefined;
@@ -8,12 +9,11 @@ export async function activate(context: vscode.ExtensionContext): Promise<void> 
   index = new MezSymbolIndex();
   await index.initialize();
 
+  const selector = { language: "heliumrapid" };
   context.subscriptions.push(
     index,
-    vscode.languages.registerDefinitionProvider(
-      { language: "heliumrapid" },
-      new MezDefinitionProvider(index)
-    )
+    vscode.languages.registerDefinitionProvider(selector, new MezDefinitionProvider(index)),
+    vscode.languages.registerReferenceProvider(selector, new MezReferenceProvider(index))
   );
 }